requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
python-dotenv
aiofiles      
PyMuPDF      
python-docx
python-multipart
